Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.1
python-dotenv==1.0.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
pytest==8.0.2
httpx==0.27.0
pytest-asyncio==0.23.5
black==24.2.0
isort==5.13.2
flake8==7.0.0 